requirements.txt
aiogram
python-dotenv
sqlalchemy
emoji